Refactor: Move getstakingstatus to rpcwallet.cpp
This is a better place to have this command, especially since it is
already categorized as a "wallet" command.
